1，完成TA的动态； 2，添加订单详情物流跟踪；

+++ MidnightArtifact/app/src/main/java/com/atman/wysq/model/request/ReportModel.java
@@ -11,9 +11,9 @@ public class ReportModel {
 
     private String content;
     private int type;
-    private int obj_id;
+    private long obj_id;
 
-    public ReportModel(String content, int type, int obj_id) {
+    public ReportModel(String content, int type, long obj_id) {
         this.content = content;
         this.obj_id = obj_id;
         this.type = type;
